refactor(project): move executors to package with same name as QActor
