+++ project/src/it/unibo/robot_adapter/demo/DemoExecutor.java
@@ -1,7 +1,7 @@
-package it.unibo.robotadapter.demo;
+package it.unibo.robot_adapter.demo;
 
 import it.unibo.qactors.akka.QActor;
-import it.unibo.robotadapter.IRobotExecutor;
+import it.unibo.robot_adapter.IRobotExecutor;
 
 /**
  * Simple RobotExecutor that only prints output on QActor standard output.
@@ -20,7 +20,7 @@ public class DemoExecutor implements IRobotExecutor {
 
     @Override
     public void doMove(final QActor qa, final String cmd) {
-        qa.println("Executing commmand " + cmd);
+        qa.println("Demo robot executing commmand " + cmd);
     }
 
 }
